refactor(ecology): make wet placement ops self-contained
