+++ mods/mod-swooper-maps/src/domain/ecology/ops/plan-wet-placement-marsh/strategies/default.ts
@@ -1,11 +1,10 @@
 import { createStrategy } from "@swooper/mapgen-core/authoring";
 
 import PlanWetPlacementMarshContract from "../contract.js";
-import { normalizeWetFeaturePlacementsConfig } from "../../../shared/wet-feature-placements/normalize-config.js";
-import { planWetFeaturePlacementsShared } from "../../../shared/wet-feature-placements/plan-wet-feature-placements-shared.js";
+import { normalizePlanWetPlacementMarshConfig } from "../rules/normalize-config.js";
+import { planWetPlacementMarsh } from "../rules/plan-wet-placement-marsh.js";
 
 export const defaultStrategy = createStrategy(PlanWetPlacementMarshContract, "default", {
-  normalize: (config) => normalizeWetFeaturePlacementsConfig(config),
-  run: (input, config) =>
-    planWetFeaturePlacementsShared({ input, config, featureKey: "FEATURE_MARSH" }),
+  normalize: (config) => normalizePlanWetPlacementMarshConfig(config),
+  run: (input, config) => planWetPlacementMarsh(input, config),
 });
