+++ mods/mod-swooper-maps/src/domain/ecology/ops/plan-wet-placement-oasis/rules/normalize-config.ts
@@ -1,0 +1,33 @@
+import { clampChance } from "@swooper/mapgen-core";
+import type { Static } from "@swooper/mapgen-core/authoring";
+
+import PlanWetPlacementOasisContract from "../contract.js";
+
+type Config = Static<(typeof PlanWetPlacementOasisContract)["strategies"]["default"]>;
+
+function normalizeRadius(value: number): number {
+  return Math.max(1, Math.floor(value));
+}
+
+export function normalizePlanWetPlacementOasisConfig(config: Config): Config {
+  const rules = config.rules;
+  return {
+    ...config,
+    multiplier: Math.max(0, config.multiplier),
+    chances: {
+      FEATURE_MARSH: clampChance(config.chances.FEATURE_MARSH),
+      FEATURE_TUNDRA_BOG: clampChance(config.chances.FEATURE_TUNDRA_BOG),
+      FEATURE_MANGROVE: clampChance(config.chances.FEATURE_MANGROVE),
+      FEATURE_OASIS: clampChance(config.chances.FEATURE_OASIS),
+      FEATURE_WATERING_HOLE: clampChance(config.chances.FEATURE_WATERING_HOLE),
+    },
+    rules: {
+      ...rules,
+      nearRiverRadius: normalizeRadius(rules.nearRiverRadius),
+      coastalAdjacencyRadius: normalizeRadius(rules.coastalAdjacencyRadius),
+      isolatedRiverRadius: normalizeRadius(rules.isolatedRiverRadius),
+      isolatedSpacingRadius: normalizeRadius(rules.isolatedSpacingRadius),
+    },
+  };
+}
+
