+++ mods/mod-swooper-maps/src/domain/ecology/ops/plan-wet-placement-tundra-bog/strategies/default.ts
@@ -1,11 +1,10 @@
 import { createStrategy } from "@swooper/mapgen-core/authoring";
 
 import PlanWetPlacementTundraBogContract from "../contract.js";
-import { normalizeWetFeaturePlacementsConfig } from "../../../shared/wet-feature-placements/normalize-config.js";
-import { planWetFeaturePlacementsShared } from "../../../shared/wet-feature-placements/plan-wet-feature-placements-shared.js";
+import { normalizePlanWetPlacementTundraBogConfig } from "../rules/normalize-config.js";
+import { planWetPlacementTundraBog } from "../rules/plan-wet-placement-tundra-bog.js";
 
 export const defaultStrategy = createStrategy(PlanWetPlacementTundraBogContract, "default", {
-  normalize: (config) => normalizeWetFeaturePlacementsConfig(config),
-  run: (input, config) =>
-    planWetFeaturePlacementsShared({ input, config, featureKey: "FEATURE_TUNDRA_BOG" }),
+  normalize: (config) => normalizePlanWetPlacementTundraBogConfig(config),
+  run: (input, config) => planWetPlacementTundraBog(input, config),
 });
