+++ mods/mod-swooper-maps/src/domain/ecology/ops/plan-wet-placement-watering-hole/rules/has-adjacent-feature-type.ts
@@ -1,0 +1,22 @@
+export function hasAdjacentFeatureType(
+  featureField: Int16Array,
+  width: number,
+  height: number,
+  x: number,
+  y: number,
+  featureIdx: number,
+  radius: number
+): boolean {
+  for (let dy = -radius; dy <= radius; dy++) {
+    const ny = y + dy;
+    if (ny < 0 || ny >= height) continue;
+    for (let dx = -radius; dx <= radius; dx++) {
+      const nx = x + dx;
+      if (nx < 0 || nx >= width) continue;
+      if (dx === 0 && dy === 0) continue;
+      if (featureField[ny * width + nx] === featureIdx) return true;
+    }
+  }
+  return false;
+}
+
